third_party/WebKit/LayoutTests/http/tests/websocket/invalid-masked-frames-from-server.html

<!DOCTYPE HTML>
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
description("Test whether WebSocket aborts the connection when it receives an masked frames from server.");

window.jsTestIsAsync = true;

var url = "ws://localhost:8880/invalid-masked-frames-from-server";
var ws = new WebSocket(url);
var closeEvent;

ws.onopen = function()
{
    debug("onopen() was called.");
};

ws.onmessage = function(event)
{
    var message = event.data;
    testFailed("onmessage() was called. (message = \"" + message + "\")");
};

ws.onclose = function(event)
{
    closeEvent = event;
    shouldBeFalse("closeEvent.wasClean");
    finishJSTest();
};

ws.onerror = function(errorEvent)
{
    testPassed("onerror() was called");
};

</script>
</body>
</html>
