third_party/blink/web_tests/http/tests/websocket/invalid-masked-frames-from-server-expected.txt

Test whether WebSocket aborts the connection when it receives an masked frames from server.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

onopen() was called.
PASS onerror() was called
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
